requirements.txt
langchain==0.0.326
openai==0.28.1
unstructured==0.10.27
faiss-cpu==1.7.4
asyncio==3.4.3
sse-starlette==1.6.5
python-dotenv==1.0.0
cohere==4.32
websocket-client==1.6.3
tls-client==0.2.2
PyPasser==0.0.5
names==0.3.0
colorama==0.4.6
curl-cffi==0.5.9
selenium==4.13.0
fake-useragent==1.3.0
chromadb==0.4.17
wikipedia==1.4.0
duckduckgo-search==3.9.3
twocaptcha==0.0.1
pyowm==3.3.0
pysrt==1.1.2
youtube-transcript-api==0.6.1
deeplake==3.8.4
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.4
filelock==3.13.0
httpx==0.25.0
Jinja2==3.1.2
jmespath==1.0.1
jsonmerge==1.9.2
jsonschema==4.19.1
Mako==1.2.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
openapi-schema-pydantic==1.2.4
pinecone-client==2.2.4
pycparser==2.21
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
qdrant-client==1.6.4
redis==4.3.4
regex==2023.5.5
replicate==0.15.4
requests==2.31.0
requests-file==1.5.1
requests-html==0.10.0
six==1.16.0
tiktoken==0.5.1
ujson==5.8.0
watchfiles==0.21.0
weaviate-client==3.25.1
yarl==1.9.2
zipp==3.17.0
pylint==3.0.2
Pillow==9.5.0
EbookLib==0.18
html2text==2020.1.16
google-generativeai==0.2.2
ai21==1.2.8
typing-extensions==4.8.0
sentence_transformers==2.2.2
g4f==0.1.7.8
gptcache==0.1.42
docker==6.1.3
pyspark==3.5.0
IPython==8.16.1
lancedb==0.3.3
text_generation==0.6.1
pymilvus==2.3.2
pydantic_settings==2.0.3
together==0.2.6
gradientai==1.2.0
librosa==0.10.1
soundfile==0.12.1
datasets==2.14.6
validators==0.22.0
novita_client==0.4.5
transformers==4.35.0 